Cleaning

+++ src/net/medox/neonengine/core/Util.java
@@ -151,7 +151,7 @@ public class Util{
 		int[] result = new int[data.length];
 		
 		for(int i = 0; i < data.length; i++){
-			result[i] = data[i].intValue();
+			result[i] = data[i];
 		}
 		
 		return result;
